rename `keychain` to `keyring`
