+++ crates/crypto/Cargo.toml
@@ -46,11 +46,11 @@ specta = { workspace = true, optional = true }
 # for asynchronous crypto
 tokio = { workspace = true, features = ["io-util", "rt-multi-thread"] }
 
-# linux OS keychain
+# linux OS keyring
 [target.'cfg(target_os = "linux")'.dependencies]
 secret-service = "2.0.2"
 
-# macos/ios OS keychain
+# macos/ios OS keyring
 [target.'cfg(any(target_os = "macos", target_os = "ios"))'.dependencies]
 security-framework = "2.8.1"
