+++ crates/crypto/src/keys/keyring/linux.rs
@@ -1,11 +1,11 @@
-//! This is Spacedrive's Linux keychain implementation, which makes use of the Secret Service API.
+//! This is Spacedrive's Linux keyring implementation, which makes use of the Secret Service API.
 //!
 //! This does strictly require `DBus`, and either `gnome-keyring`, `kwallet` or another implementor of the Secret Service API.
 
 use secret_service::{Collection, EncryptionType, SecretService};
 
 use crate::{
-	keys::keychain::{Identifier, Keyring},
+	keys::keyring::{Identifier, Keyring},
 	Error, Protected, Result,
 };
