+++ crates/crypto/src/keys/mod.rs
@@ -1,5 +1,5 @@
 //! This module contains all key and hashing related functions.
 
 pub mod hashing;
-pub mod keychain;
 pub mod keymanager;
+pub mod keyring;
